Added G3NetworkSender header file
